refactor(query-plan): split apply_joins into cohesive planner stages
Extract the 1,586-line apply_joins implementation into focused sibling
modules without changing join planning behavior:

- selection: collect nested selects, synthetic ORDER BY joins, join info
- child_mapping: enrich scan mapping for aggregates/filters/order/BM25
- child_plan: resolve target collection and build child scan plans
- sub_joins: recursive joins, order/agg/multi-level/filter sub-joins, ACP
- selection_join: per-relation pipeline orchestrator
- join_many / join_one: TypeJoinMany and TypeJoinOne assembly
- mod.rs: thin apply_joins orchestrator (~140 lines)

Preserves branch order, error strings, tracing, and external module paths.
Production type_join_one.rs is intentionally out of scope.
